refactor(ui): move supply request list into separate component
