+++ src/utils/api-types.ts
@@ -1,4 +1,4 @@
-import { SupplyRequestStatus } from '.prisma/client';
+import { SupplyRequestStatus } from '@prisma/client';
 
 type SupplyRequestItem = {
   id: number;
